Reorganize Config.cs into Terraria and ModLoader refactors
